+++ Routes/manager_routes.py
